Проверка и написание couriers/complete

+++ store/models/completed_order.py
@@ -5,5 +5,5 @@ class CompletedOrder(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     courier_id = db.Column(db.Integer, db.ForeignKey('courier.courier_id'))
     completed_orders = db.Column(db.Integer, default=0)
-    min_time = db.Column(db.Integer)
+    min_time = db.Column(db.Float)
     last_complete_time = db.Column(db.DateTime)
